Refactor mpris plugin
